requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1.1
djangorestframework==3.11.1
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
